dedicated file for CRUD operations regarding the menu logic
